Allow handling of multiple datasets
Now multiple datasets may be loaded concurrently. All the data and the METADATA
file associated with a dataset must be in a dedicated folder within the main
data folder.

Prior builds may break, the datafiles will need to be moved into a subfolder. I
will fix this in the future.

The datasets are entirely independent. I may add a way to link between them,
but I have no immediate plans to this end.