Add internal KEY and functional id selection
Previously the 'Enter ids' box that allowed users to select rows by
pasting in lists of ids only worked for the key column. This key column
was taken from the first column of the dataset. In this commit, I
removed the special status of that key column and created a new internal
key. The 'Enter ids' box now works for any key-like column (it doesn't
work for textual or sequence columns).