Fix error on first access of cor column
An error occurs when Matrix is autoloaded the first time a cor column is
plotted. I can avoid this problem by loading matrix when the app
initiates. This is not an optimal solution, since Matrix is a large
program and only is needed if there are textual columns in the dataset.