Fix crash when submitting user keys
The server.R::setFilter and server.R::user.keys functions are
abominations. I have had no end of trouble with user.keys. DT handles
the rows that are displayed based on the column filters. When the user
searches for keys, I simply subset the whole dataset prior to DT getting
it. Thus, DT is ignorant of the full dataset. This means the total rows
it displays is false. This also means column_table_rows_selected does
not correspond directly to row numbers in the dataset. All this is a
little awkward and messy. I have found the whole thing terribly fragile
as well. I need to nail it down with rigorous tests.