Add handling of volatile input in Sandbox mode
My implementation may not be safe as is. I haven't tested its behavior
when mutliple users log in simultaneously. But I believe the data a user
uploads will be visible to someone who logs in after they do.
Furthermore, when the user logs off, both their data and all other users
data will be erased. This needs to be fixed. But for now, I would just
caution users to upload sensitive data.