Better handling of instance specific files
Different Metaoku instances will have different configurations (config
file) and potentially different information in the home page, about
page, etc. Previously I modified the local copies of the
version-controlled config and the pages.

I now separate the version controlled defaults into the inst/ directory,
and make local copies with a Makefile. This allows the user to

 1. initialize a new instance with a simple `make` call

 2. clean up the repository with `make clean`, (instance-specific
 material will be backed-up in `/tmp`).

 3. pull the most recent versions of Metaoku without overwriting or
 conflicting with instance-specific modifications

 4. refer to the unmodified default configurations

Also, all instance-specific material is specified in the .gitignore
file, so it won't be accidentally committed.